--- c23c_601_2/src/main/java/com/c23c_601_2/dao/WebProjectDAO.java
@@ -1,5 +1,0 @@
-package com.c23c_601_2.dao;
-
-public class WebProjectDAO {
-
-}
